Removed `"1"` in `{{#filterList operationParams "op=>op.inBody" "1"}}
It seemed like the `"1"` in the filter nade the iteration stop afeter one iteration. That led to the code not running if the body parameter wasn't the first parameter.

I hadn't time to figure out exactly how the filterList helper actually works but for my intents and purposed the fix works. It migh break something else though.